added new headers to accommodate core data
